backend/pyproject.toml
# backend/pyproject.toml
[project]
name = "travel-agent-pro"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agent*", "llm*", "state*", "phase*", "tools*", "context*", "memory*", "harness*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
